+++ .gitignore
@@ -13,7 +13,7 @@
 # testing
 /coverage
 
-# next.js
+# next.js 
 /.next/
 /out/
