+++ src/hooks/Article/index.js
@@ -1,0 +1,5 @@
+// 임포트 경로 단축용
+
+export { useArticles } from "./useArticles";
+export { useArticle } from "./useArticle";
+export { useComments } from "./useComments";
